README
dBoard is a smart contract that can be used to create motions and do a vote using the ethereum blockchain.

While building this smart contract we learned a lot about deploying a smart contract on the ethereum blockchain and buiding an enviornment and running a node with truffle. We also faced the complications of bringing together a front end and backend.